refactor: break down cache eviction.rs into focused modules
- Split 457-line eviction.rs into organized module hierarchy
- All files under 200 lines (largest: 176 lines)
- Group by function: traits, factory, policies/, tests
- Use directory hierarchies, no underscores
- Maintain all functionality and tests
