+++ crates/cache/src/eviction/policies/mod.rs
@@ -1,0 +1,9 @@
+//! Eviction policy implementations
+
+mod arc;
+mod lfu;
+mod lru;
+
+pub use arc::ArcPolicy;
+pub use lfu::LfuPolicy;
+pub use lru::LruPolicy;
